requirements.txt
streamlit
langchain
langchain-openai
langchain-community
chromadb
pydantic
python-dotenv
tenacity
requests
beautifulsoup4
schedule